requirements.txt
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2
django-classy-tags==0.9.0
django-import-export==2.0
django-sekizai==1.0.0
django-session-timeout==0.0.4
et-xmlfile==1.0.1
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.3
Pillow==7.0.0
pytz==2019.3
PyYAML==5.3
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
xlrd==1.2.0
xlwt==1.3.0
